Move methods to building.cpp
